+++ site-modules/role/spec/classes/baseline_spec.rb
@@ -3,20 +3,17 @@
 require 'spec_helper'
 
 describe 'role::baseline' do
-  let(:facts) do
-    {
-      os: {
-        'family' => 'Debian',
-        'name' => 'Debian',
-        'release' => { 'major' => '12' }
-      }
-    }
-  end
+  SUPPORTED_FACTS.each do |platform, platform_facts|
+    context "on #{platform}" do
+      let(:facts) { platform_facts }
+      let(:trusted_facts) do
+        { 'certname' => platform_facts[:networking]['fqdn'], 'authenticated' => 'local', 'extensions' => {} }
+      end
 
-  it { is_expected.to compile.with_all_deps }
-  it { is_expected.to contain_class('profile::baseline') }
-
-  it 'does not manage workload resources in Milestone 1' do
-    expect(non_structural_resources(catalog)).to be_empty
+      it { is_expected.to compile.with_all_deps }
+      it { is_expected.to contain_class('profile::baseline') }
+      it { is_expected.to contain_package('git').with_ensure('installed') }
+      it { is_expected.to contain_file('/etc/sasd/puppet-baseline.conf') }
+    end
   end
 end